requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
numpy
pandas
sentence-transformers
torch
rank-bm25
pymupdf
python-docx
python-pptx
pillow